Check for an empty /etc/os-release
An empty /etc/os-release file is a special case for the mock build
environment.

In a real-world install, /etc/os-release is a symlink to
`../usr/lib/os-release`, which works correctly, since os.path.isfile()
follows the symlink.

Note: This change fixes the build for WALinuxAgent, which probes for the
distro details to know how/where to install content.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>